+++ tests/test_apps/genqa/src/genqa/RoughCSVTokenizer.java
@@ -20,7 +20,7 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
  * OTHER DEALINGS IN THE SOFTWARE.
  */
-package testutils;
+package genqa;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
