--- tests/test_apps/testutils/.gitignore
@@ -1,1 +1,0 @@
-/bin/
