--- tests/test_apps/testutils/build.xml
@@ -1,39 +1,0 @@
-<!-- Simple build file to build test -->
-<project name="testutils" basedir="." default="all">
-
-    <property name="build.dir"   value="bin"/>
-    <property name="jar.dir"     value="${basedir}"/>
-    <property name="src.dir"     value="src"/>
-    <property name="voltdbroot.dir"     value="${basedir}/voltdbroot"/>
-    <property name="log.dir"     value="${basedir}/log"/>
-
-    <target name="all" depends="testutils"/>
-
-    <target name="clean">
-        <delete dir="${build.dir}"/>
-    </target>
-
-    <path id="build-classpath">
-       <fileset dir="${basedir}/../../../voltdb">
-         <include name="voltdb*.jar"/>
-       </fileset>
-       <fileset dir="${basedir}/../../../lib">
-           <include name="*.jar"/>
-       </fileset>
-    </path>
-
-    <target name="testutils-compile">
-        <mkdir dir="${build.dir}"/>
-        <javac debug="on" srcdir="${src.dir}"
-            destdir="${build.dir}" >
-            <classpath refid="build-classpath"/>
-        </javac>
-    </target>
-
-    <target name="testutils" depends="testutils-compile">
-        <mkdir dir="${jar.dir}"/>
-        <jar destfile="${jar.dir}/testutils.jar" basedir="${build.dir}">
-        </jar>
-    </target>
-
-</project>
